# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ubademy - Backend",
    "description": "Ubademy -  Coursera like app, for taller2",
    "license": {
      "name": "FIUBA",
      "url": "https://fi.uba.ar/"
    }
  },
  "host": "ubademy-la-venganza-back.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/course/types": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Search courses' types",
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/course": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Create new course",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "The user to create.",
            "schema": {
              "$ref": "#/definitions/course"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Search courses",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "in": "query",
            "name": "keyword",
            "type": "string"
          },
          {
            "in": "query",
            "name": "plan",
            "type": "string"
          },
          {
            "in": "query",
            "name": "category",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/course/{course_id}": {
      "parameters": [
        {
          "in": "path",
          "name": "course_id",
          "type": "integer",
          "required": true,
          "description": "Numeric ID of the course to get"
        }
      ],
      "get": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "description": "ID of the user requesting the course"
          }
        ],
        "summary": "Get course's content (Information is different depending user type)",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Courses"
        ],
        "summary": "Modify course's content",
        "responses": {
          "202": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/course/{id}/registration": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "type": "integer",
          "required": true,
          "description": "Numeric ID of the course"
        }
      ],
      "post": {
        "tags": [
          "Courses"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "user_id"
              ],
              "properties": {
                "user_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "summary": "Register as a student to course",
        "responses": {
          "201": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Courses"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "user_id"
              ],
              "properties": {
                "user_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "summary": "Unregister as a student to course",
        "responses": {
          "201": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
    "/exam/course/{course_id}": {
      "get": {
        "tags": [
          "Exams"
        ],
        "summary": "Staff search exams basic info for course by id",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "type": "integer",
            "required": true,
            "description": "ID for the course to get exam from"
          },
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "course": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "hashtags": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "lessons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/lesson"
          }
        }
      }
    },
    "lesson": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "require": {
          "type": "boolean"
        },
        "acive": {
          "type": "boolean"
        },
        "position": {
          "type": "integer"
        },
        "multimedia_id": {
          "type": "string"
        },
        "multimedia_type": {
          "type": "string"
        },
        "exam": {
          "$ref": "#/definitions/exam"
        }
      }
    },
    "exam": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "minimum_qualification": {
          "type": "integer"
        },
        "questions": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "sequence_number": {
                "type": "integer"
              },
              "type": {
                "type": "string"
              },
              "score": {
                "type": "integer"
              },
              "multiple_choice_question": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "amount_of_options": {
                    "type": "integer"
                  },
                  "choices": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "text": {
                          "type": "string"
                        },
                        "is_correct": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              },
              "develop_question": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}